requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# ── Auth / JWT ─────────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1
bcrypt==4.1.3

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Storage ────────────────────────────────────────────────────────────────────
boto3==1.34.84
django-storages==1.14.2

# ── Email ──────────────────────────────────────────────────────────────────────
sendgrid==6.11.0

# ── Firebase / Push ────────────────────────────────────────────────────────────
firebase-admin==6.5.0

# ── Async / Cache ──────────────────────────────────────────────────────────────
redis==5.0.4
django-redis==5.4.0

# ── Background Tasks ───────────────────────────────────────────────────────────
celery==5.3.6
django-celery-beat==2.6.0

# ── Filtering / Search ─────────────────────────────────────────────────────────
django-filter==24.2

# ── Environment ────────────────────────────────────────────────────────────────
python-decouple==3.8
dj-database-url==2.1.0

# ── Utilities ──────────────────────────────────────────────────────────────────
Pillow==10.3.0
python-magic==0.4.27